Move BufferSet* into common
